lib/goth/schema.rb,bin/goth - don't hardwire translations file
Primarily, it can't be relied on to exist in the current working
directory (we can be invoked from anywhere, and in particular
generally not in the Goth project directory).

But secondly, we may in general want to supply a translations from
outside.

In this first step, we require the translations to be supplied from
outside. Later we will embed some default translations.