bin/goth,lib/goth.rb etc. - package the default translations and templates in the Gem
i.e. Include these files in the lib/goth directory tree so that they
get installed with the Gem source, and create references to them in
goth.rb.

Use these references as defaults in bin/goth. The script will now be
able to use them whenever it is installed normally.